spit test in sperate classes
